requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.34.82
botocore==1.34.162
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.1.0
msgpack==1.1.1
openpyxl==3.1.5
pillow==10.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==6.4.0
s3transfer==0.10.4
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
zope.interface==7.2
